refactor: reduce dom.js fan-in by redirecting imports to specialized modules (closes #203)
dom-dialogs.js was acting as a facade that re-exported all dom.js symbols,
creating a 35+ module hub. Now:

- Modules needing only core DOM primitives (_el, createActionButton,
  renderButtonBar, buildChevronRow, etc.) import directly from dom.js
- dom-dialogs.js no longer re-exports dom.js primitives; it only exports
  dialog builders (showPromptDialog, showConfirmDialog, createCustomModal)
- form-helpers.js no longer re-exports _el; consumers import it from dom.js
- dom-dialogs.js fan-in: 35 → 3 (only modules that actually use dialogs)
