+++ src/utils/dom-dialogs.js
@@ -1,17 +1,13 @@
 /**
  * Dialog and prompt helpers — extracted from dom.js to reduce file length.
  *
- * Also serves as a facade: re-exports every public symbol from dom.js so that
- * consumer modules can import from this single module instead of coupling
- * directly to dom.js.  (See issue #203.)
+ * This module provides high-level dialog builders (prompt, confirm, custom
+ * modal).  For core DOM primitives (_el, createActionButton, renderButtonBar,
+ * buildChevronRow, etc.) import directly from './dom.js'.
  */
 
 import { _el, createActionButton, createModalOverlay } from './dom.js';
 import { setupKeyboardShortcuts } from './keyboard-helpers.js';
-
-// Re-export all dom.js public symbols so consumers can avoid a direct
-// dependency on the dom.js hub module.
-export { _el, createActionButton, createModalOverlay, renderButtonBar, buildChevronRow, _safeFit, createSelect, positionInViewport } from './dom.js';
 
 // ── Private dialog lifecycle ──
